Fixes error in field calculation
Multiple improvements made for the function plotting field in the
coating as a function of depth.

1. As was raised in #1,
https://git.ligo.org/40m/Coatings/-/issues/1#note_227168,
the absorption in the optimized coating seemed high. This was actually
an error in the way the fieldDepth function was being called, the
default wavelength of 1064nm was being used instead of the 2um that is
appropriate for the coating discussed in the comment. Fixed here.

2. Adds some diagnostic info to the field plots about the absorption,
for better debugging.

3. Saves the PDFs with better names.